requirements.txt
customtkinter
requests
langchain-community
langchain-ollama
langchain-huggingface
langchain-chroma
chromadb
sentence-transformers
spacy
pyspellchecker
Pillow
pandas
beautifulsoup4
markdown
python-docx
PyPDF2
openpyxl
googlesearch-python
datasets
